feat(wgpu, pianoroll, arrangement, egui): NoteInstance track 从 u8 提升到 u16，支持 65536 轨道
- NoteInstance::pack 签名改为 (key: u8, track: u16, vel: u8)，丢掉永远为 0 的 flags
- 新位布局：key(0..8) | track(8..24) | vel(24..32)，instance 仍 16B
- MAX_TRACKS: 256 → 65536（Rust + WGSL 同步）
- track_colors 改 storage buffer（1MB 超 uniform 上限）
- shader track 解位 & 0xFFu → & 0xFFFFu
- 移除所有 note.track.min(255) as u8 的 saturate 操作
- ghost_notes 类型从 (f64, f64, u8, u8) 改为 (f64, f64, u8, u16)
- 测试断言同步更新到新位布局

+++ crates/yinhe-pianoroll/src/pianoroll_prepare.rs
@@ -33,7 +33,7 @@ pub fn prepare(
     scroll_mode: u32,
     min_border_width: f32,
     midi_version: u64,
-    ghost_notes: &[(f64, f64, u8, u8)], // (start_tick, end_tick, key, track) for pencil preview
+    ghost_notes: &[(f64, f64, u8, u16)], // (start_tick, end_tick, key, track) for pencil preview
     note_selection_highlight: bool,
 ) -> yinhe_wgpu::PrepareTimings {
     let w = width as f32;
